[PIDM-1954] extracted performance tracing log logic and disabled it by default
